refactor(userInfo): userInfo 프레임 분리
ranking에 있던 userInfoFrame(avatar, intraname, username) components 폴더로 분리
