+++ components/ranking/userFrame.tsx
@@ -2,6 +2,7 @@ import styled from "styled-components";
 import Image from "next/image";
 import { useState, useEffect } from "react";
 import { bronze, silver, gold, platinum, diamond } from "./tier";
+import UserInfo from "../userInfo";
 
 const UserFrame: React.FC<{
   ranking: number;
@@ -49,18 +50,12 @@ const UserFrame: React.FC<{
           width={30}
           height={30}
         />
-        <UserInfoFrame>
-          <ProfileImage
-            src={avatar}
-            alt="Profle Image"
-            width={100}
-            height={100}
-          />
-          <NameFrame>
-            <NickNameFrame> {nickName} </NickNameFrame>
-            <IntraNameFrame> {intraName} </IntraNameFrame>
-          </NameFrame>
-        </UserInfoFrame>
+        <UserInfo
+          nickName={nickName}
+          intraName={intraName}
+          avatar={avatar}
+          size={50}
+        />
       </LeftFrame>
       <MatchCountFrame> 경기 횟수: {rankGameTotalCount}회</MatchCountFrame>
     </UserRankFrame>
@@ -102,43 +97,6 @@ const RankingFrame = styled.div`
 const TierImage = styled(Image)`
   width: 10%;
   height: auto;
-`;
-
-const UserInfoFrame = styled.div`
-  width: 50%;
-  height: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 7%;
-`;
-
-const ProfileImage = styled(Image)`
-  width: 15%;
-  height: auto;
-  border-radius: 50px;
-`;
-
-const NameFrame = styled.div`
-  width: 70%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-family: "GiantsLight";
-  color: ${(props) => props.theme.colors.brown};
-`;
-
-const NickNameFrame = styled.div`
-  width: 100%;
-  font-size: small;
-`;
-
-const IntraNameFrame = styled.div`
-  width: 100%;
-  font-size: x-small;
-  margin-top: 3%;
 `;
 
 const MatchCountFrame = styled.div`
